+++ TILApp/Models/AuthUser.swift
@@ -9,6 +9,7 @@ struct AuthUser: Codable {
     let followings: Int
     let hasBlog: Bool
     let lastPublishedAt: Date?
+    let isAgreed: Bool?
 }
 
 struct SignInInput: Codable {
